Add file previews
